sql: organize the planning constructors in separate files
The following constructors perform all the execution work in the
constructor itself, and produce a `zeroNode`:

| Constructor              | File                         | Previously in     |
|--------------------------|------------------------------|-------------------|
| `BeginTransaction`       | `begin_transaction.go`       | `txn.go`          |
| `Deallocate`             | `deallocate.go`              | `prepare.go`      |
| `Discard`                | `discard.go`                 |                   |
| `Grant`                  | `grant_revoke.go`            | `grant.go`        |
| `RenameColumn`           | `rename_column.go`           | `rename.go`       |
| `RenameDatabase`         | `rename_database.go`         | `rename.go`       |
| `RenameIndex`            | `rename_index.go`            | `rename.go`       |
| `RenameTable`            | `rename_table.go`            | `rename.go`       |
| `Revoke`                 | `grant_revoke.go`            | `grant.go`        |
| `SetDefaultIsolation`    | `set_default_isolation.go`   |                   |
| `SetTransaction`         | `set_transaction.go`         | `txn.go`          |
| `ShowGrants`             | `show_grants.go`             | `show.go`         |

The following constructors delegate their work to a SQL query,
expanded and executed via `delayedNode`:

| Constructor              | File                         | Previously in     |
|--------------------------|------------------------------|-------------------|
| `ShowClusterSetting`     | `show_cluster_setting.go`    | `show.go`         |
| `ShowColumns`            | `show_columns.go`            | `show.go`         |
| `ShowConstraints`        | `show_constraints.go`        | `show.go`         |
| `ShowCreateTable`        | `show_create.go`             | `show.go`         |
| `ShowCreateView`         | `show_create.go`             | `show.go`         |
| `ShowDatabases`          | `show_databases.go`          | `show.go`         |
| `ShowIndex`              | `show_index.go`              | `show.go`         |
| `ShowJobs`               | `show_jobs.go`               | `show.go`         |
| `ShowQueries`            | `show_queries.go`            | `show.go`         |
| `ShowSessions`           | `show_sessions.go`           | `show.go`         |
| `ShowTables`             | `show_tables.go`             | `show.go`         |
| `ShowTransactionStatus`  | `show_transaction_status.go` | `show.go`         |
| `ShowUsers`              | `show_users.go`              | `show.go`         |
| `ShowVar`                | `show_var.go`                | `show.go`         |

The remaining constructors operate on a "native" planNode. Only a few
are relocated in this patch:

| Constructor              | Plan node                       | File                         | Previously in         |
|--------------------------|---------------------------------|------------------------------|-----------------------|
| `AlterSequence`          | `alterSequence`                 | `alter_sequence.go`          |                       |
| `AlterTable`             | `alterTable`                    | `alter_table.go`             |                       |
| `AlterUserSetPassword`   | `alterUserSetPassword`          | `alter_user.go`              |                       |
| `CancelJob`              | `controlJob`                    | `control_job.go`             |                       |
| `CancelQuery`            | `cancelQuery`                   | `cancel_query.go`            |                       |
| `CopyFrom`               | `copy`                          | `copy.go`                    |                       |
| `CreateDatabase`         | `createDatabase`                | `create_database.go`         |                       |
| `CreateIndex`            | `createIndex`                   | `create_index.go`            |                       |
| `CreateSequence`         | `createSequence`                | `create_sequence.go`         |                       |
| `CreateTable`            | `createTable`                   | `create_table.go`            |                       |
| `CreateUser`             | `createUser`                    | `create_user.go`             |                       |
| `CreateView`             | `createView`                    | `create_view.go`             |                       |
| `Delete`                 | `delete`                        | `delete.go`                  |                       |
| `DropDatabase`           | `dropDatabase`                  | `drop_database.go`           |                       |
| `DropIndex`              | `dropIndex`                     | `drop_index.go`              |                       |
| `DropSequence`           | `dropSequence`                  | `drop_sequence.go`           |                       |
| `DropTable`              | `dropTable`                     | `drop_table.go`              |                       |
| `DropUser`               | `dropUser`                      | `drop_user.go`               |                       |
| `DropView`               | `dropView`                      | `drop_view.go`               |                       |
| `Execute`                | (prepared)                      | `execute.go`                 | `prepare.go`          |
| `Explain`                | `explainPlan`, `explainDistSQL` | `explain.go`                 |                       |
| `Insert`                 | `insert`                        | `insert.go`                  |                       |
| `PauseJob`               | `controlJob`                    | `control_job.go`             |                       |
| `ResumeJob`              | `controlJob`                    | `control_job.go`             |                       |
| `Scatter`                | `scatter`                       | `scatter.go`                 |                       |
| `Scrub`                  | `scrub`                         | `scrub.go`                   |                       |
| `Select`, `SelectClause` | `render`                        | `render.go`                  |                       |
| `SetClusterSetting`      | `setClusterSetting`             | `set_cluster_setting.go`     |                       |
| `SetVar`                 | `set`                           | `set_var.go`                 | `set.go`              |
| `SetZoneConfig`          | `setZoneConfig`                 | `set_zone_config.go`         |                       |
| `ShowFingerprints`       | `showFingerprints`              | `show_fingerprints.go`       |                       |
| `ShowRanges`             | `showRanges`                    | `show_ranges.go`             |                       |
| `ShowTrace`              | `trace`                         | `show_trace.go`              | `show.go`, `trace.go` |
| `ShowZoneConfig`         | `showZoneConfig`                | `show_zone_config.go`        |                       |
| `Split`                  | `split`                         | `split.go`                   |                       |
| `TestingRelocate`        | `testingRelocate`               | `testing_relocate.go`        |                       |
| `UnionClause`            | `union`                         | `union.go`                   |                       |
| `Update`                 | `update`                        | `update.go`                  |                       |
| `ValuesClause`           | `values`                        | `values.go`                  |                       |
